backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[dotenv]
sqlalchemy
psycopg2-binary
web3
eth-account
python-dotenv
geopy
